It works! But the stalls are still slightly off

+++ trace_decode_header.h
@@ -21,12 +21,12 @@ class Instruction_decoder
 		int reg_rs = 0;
 		int reg_rt = 0;
 		int reg_rd = 0;
-		int immediate = 0;
+		long long int immediate = 0;
 		int source1 = 0;		// Source register to be used by pipeline
 		int source2 = 0;		// Second source register to be used by pipeline
 		int destination = 0;	// Destination register to be used by pipeline
 		int x_addr = 0;			// Address in memory
-		int entire_value = 0;	// Unaltered read in value
+		long long int entire_value = 0;	// Unaltered read in value
 		bool branch_taken = false;	
 		string type = "U";
